GUI done, Main done (but will need bug fixes)
GUI should be final. Main is done, but it is awaiting all the functionalitity from the game and computer player class. Once those are done, any bugs in it can be worked out. If you want to get the program to run, comment out the part were its says "# game begins here!"